feat: Add "b" script to package.json for building the exporter
This commit adds a new script "b" to the package.json file. The script is used for building the exporter by running the "bun" command with specific options. This change improves the build process by providing a convenient way to compile and minify the exporter source code and output it to the dist/exporter directory.